Backend/requirements.txt
ipykernel
fastapi
uvicorn
python-multipart
PyPDF2
python-docx
nltk
openai
python-dotenv
sentence-transformers
numpy